requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Logging and experiment tracking
wandb>=0.15.0

# Data handling
pyyaml>=6.0
numpy>=1.24.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pandas>=2.0.0
python-dotenv>=1.0.0
roboflow>=1.0.0